chore: update invite sheet to deal with double render
